// src/models/user.ts
export interface User {
  id: string | "";
  name: string;
  lastname: string;
  dni: string;
  mail: string;
  user: string | "";
  password: string | "";
  rol: string | "";
  birth: string | "";
  address: string | "";
  phone: string | "";
  vaccinated: boolean | "";
  typeOfVaccine: string | "";
  date: string | "";
  dosis: number | "";
}